+++ src/routes/components/+page.svelte
@@ -1,26 +1,6 @@
 <script>
-	import { persist, localStorage } from '@macfja/svelte-persistent-store';
-	import { onMount } from 'svelte';
-	import { writable } from 'svelte/store';
-
 	import components from './components.json';
-	import List from '$components/ComponentIndex/CardList.svelte';
-	import Select from '$components/Select.svelte';
-	import SearchLayout from '$layouts/SearchLayout.svelte';
-	import ComponentCard from '$lib/components/ComponentIndex/Card.svelte';
-	import { extractUnique } from '$lib/utils/extractUnique';
-	import Seo from '$lib/components/Seo.svelte';
-	import Search from '$lib/components/Search.svelte';
-
-	let searchValue;
-	let selectedPackageManager = { value: 'npm' };
-	let packageManager = writable('npm');
-	let dataToDisplay = [];
-	onMount(() => {
-		packageManager = persist(writable('npm'), localStorage(), 'packageManager');
-	});
-	$: $packageManager = selectedPackageManager.value;
-	$: categories = extractUnique(dataToDisplay, 'category');
+	import SearchableJson from '../_searchableJson.svelte';
 
 	const categoryId = {
 		Animations: 'animations',
@@ -39,74 +19,10 @@
 	};
 </script>
 
-<Seo title="Components" />
-
-<SearchLayout title="Components">
-	<section class="controls" slot="controls">
-		<div class="inputs">
-			<Search
-				data={components}
-				dataDefault={{ category: '' }}
-				bind:query={searchValue}
-				sortableFields={[
-					{ identifier: 'addedOn', title: 'Last added first', ascending: false },
-					{ identifier: 'addedOn', title: 'Oldest first', ascending: true },
-					{ identifier: 'title', title: 'Name (A-Z)', ascending: true },
-					{ identifier: 'title', title: 'Name (Z-A)', ascending: false },
-					{ identifier: 'stars', title: 'Most star first', ascending: false }
-				]}
-				searchableFields={['title', 'description']}
-				facetsConfig={[
-					{
-						title: 'Category',
-						identifier: 'category',
-						isClearable: true,
-						showIndicator: true
-					},
-					{
-						title: 'Tags',
-						identifier: 'tags',
-						isMulti: true
-					}
-				]}
-				on:search={(a) => (dataToDisplay = a.detail.data.items)}
-			/>
-			<Select
-				label="Package manager"
-				isClearable={false}
-				showIndicator
-				bind:value={selectedPackageManager}
-				items={[
-					{ label: 'NPM', value: 'npm' },
-					{ label: 'PNPM', value: 'pnpm' },
-					{ label: 'Yarn', value: 'yarn' }
-				]}
-			/>
-		</div>
-
-		<a href="/help/submitting?type=component" class="submit">Submit a component</a>
-		<input
-			class="searchbar"
-			type="text"
-			placeholder="Search for components..."
-			bind:value={searchValue}
-		/>
-		<span class="searchbar-count"
-			>{dataToDisplay.length} result{#if dataToDisplay.length !== 1}s{/if}</span
-		>
-	</section>
-	<section slot="items">
-		{#each categories as category}
-			<List
-				title={category.label || 'Unclassified'}
-				id={categoryId[category.label] || category.label || 'unclassified'}
-			>
-				{#each dataToDisplay.filter((d) => d.category === category.value || (!categories
-							.map((v) => v.value)
-							.includes(d.category) && category.value === '')) as data}
-					<ComponentCard {...data} manager={$packageManager} />
-				{/each}
-			</List>
-		{/each}
-	</section>
-</SearchLayout>
+<SearchableJson
+	{categoryId}
+	data={components}
+	displayTitle="Components"
+	displayTitleSingular="component"
+	submittingType="component"
+/>
